+++ jaraco/tidelift/__init__.py
@@ -1,0 +1,5 @@
+from . import banner
+
+
+def setup(app):
+    banner.setup(app)